+++ src/BuiltWithBadge/index.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import './builtwith.css'
 
-export default function BuiltWith({ href }) {
+export default function BuiltWith({ href, ...props }) {
   return (
-    <a href={href} className="formkit-powered-by-convertkit" target="_blank" rel="noopener noreferrer" data-element="powered-by">
+    <a href={href} className="formkit-powered-by-convertkit" target="_blank" rel="noopener noreferrer" data-element="powered-by" {...props}>
       Built with ConvertKit
     </a>
   )